+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ActionUiHelper.kt
@@ -690,6 +690,8 @@ class ActionUiHelper(
 
         ActionData.ToggleKeyboard -> getString(R.string.action_toggle_keyboard)
 
+        ActionData.CycleKeyboardLanguage -> getString(R.string.action_cycle_keyboard_language)
+
         ActionData.ToggleSplitScreen -> getString(R.string.action_toggle_split_screen)
 
         ActionData.VoiceAssistant -> getString(R.string.action_open_assistant)