+++ base/src/test/java/io/github/sds100/keymapper/base/system/inputmethod/FakeInputMethodAdapter.kt
@@ -36,4 +36,8 @@ class FakeInputMethodAdapter : InputMethodAdapter {
             ?.let { Success(it) }
             ?: KMError.InputMethodNotFound(packageName)
     }
+
+    override fun cycleInputMethodSubtype(): KMResult<Unit> {
+        return Success(Unit)
+    }
 }